Remove diff viewer
