--- .github/workflows/deploy-preview.yml
@@ -1,36 +1,0 @@
-name: "Publish Preview Deployment"
-
-on:
-    workflow_run:
-        workflows: ["Build Preview Deployment"]
-        types:
-            - completed
-
-permissions:
-    actions: read
-    contents: read
-    deployments: write
-    pull-requests: write
-
-jobs:
-    deploy-preview:
-        runs-on: "ubuntu-latest"
-        if: ${{ github.event.workflow_run.conclusion == 'success' }}
-        name: "Deploy Preview to Cloudflare Pages"
-        steps:
-            - name: "Download build artifact"
-              uses: "actions/download-artifact@v4"
-              id: "preview-build-artifact"
-              with:
-                  name: "preview-build"
-                  path: "dist"
-                  github-token: ${{ secrets.GITHUB_TOKEN }}
-                  run-id: ${{ github.event.workflow_run.id }}
-            - name: "Deploy to Cloudflare Pages"
-              uses: "AdrianGonz97/refined-cf-pages-action@v1"
-              with:
-                  apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-                  accountId: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-                  githubToken: ${{ secrets.GITHUB_TOKEN }}
-                  projectName: "patch-roulette"
-                  directory: ${{ steps.preview-build-artifact.outputs.download-path }}
