+++ .github/workflows/test.yml
@@ -27,5 +27,3 @@ jobs:
       - run: cd web && bun install
       - run: cd web && bun run lint
       - run: cd web && bun run check
-      - run: cd web-extension && bun install
-      - run: cd web-extension && bun run lint
