--- web-extension/.prettierrc
@@ -1,9 +1,0 @@
-{
-    "endOfLine": "lf",
-    "useTabs": false,
-    "tabWidth": 4,
-    "printWidth": 160,
-    "plugins": ["prettier-plugin-tailwindcss"],
-    "tailwindStylesheet": "./src/style.css",
-    "overrides": []
-}
