--- web-extension/README.md
@@ -1,9 +1,0 @@
-# Patch Roulette Web Extension
-
-## Building
-
-1. Install the latest version of [`bun`](https://bun.sh/) (1.2.9 at time of writing)
-2. Run `bun install` to install dependencies
-3. Run `bun run pack` to build the extension
-
-The built extensions will be in the `pack` directory.
