--- web-extension/options.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <link rel="icon" type="image" href="/favicon.png" />
-        <title>Patch Roulette Extension Options</title>
-        <script type="module" src="/src/options.ts"></script>
-        <link rel="stylesheet" href="/src/style.css" />
-    </head>
-    <body class="p-3 text-base">
-        <h1 class="mb-4 text-3xl">Patch Roulette Extension Options</h1>
-        <label class="flex flex-col">
-            <span class="text-lg">Viewer URL</span>
-            <input id="url-input" type="text" class="rounded border border-gray-300 px-2 py-1 overflow-ellipsis" />
-        </label>
-    </body>
-</html>
