--- web-extension/package.json
@@ -1,43 +1,0 @@
-{
-    "name": "patch-roulette-web-extension",
-    "displayName": "Patch Roulette",
-    "description": "Streamlines opening diffs in the Patch Roulette Diff Viewer.",
-    "version": "0.0.6",
-    "private": true,
-    "type": "module",
-    "scripts": {
-        "format": "prettier --write .",
-        "lint": "tsc && prettier --check . && eslint .",
-        "manifest:chrome": "bun scripts/manifest.ts",
-        "manifest:firefox": "EXTENSION=firefox bun scripts/manifest.ts",
-        "build": "vite build",
-        "pack:chrome": "bun run manifest:chrome && bun run build && mkdir -p pack && rm -f pack/chrome.zip && (cd dist && zip -r ../pack/chrome.zip .)",
-        "pack:firefox": "bun run manifest:firefox && bun run build && mkdir -p pack && rm -f pack/firefox.zip && (cd dist && zip -r ../pack/firefox.zip .)",
-        "pack": "bun run pack:chrome && bun run pack:firefox"
-    },
-    "devDependencies": {
-        "@eslint/compat": "^1.2.8",
-        "@eslint/js": "^9.25.1",
-        "@types/bun": "^1.2.10",
-        "@types/webextension-polyfill": "^0.12.3",
-        "chrome-types": "^0.1.348",
-        "eslint": "^9.25.1",
-        "eslint-config-prettier": "^10.1.2",
-        "globals": "^16.0.0",
-        "prettier": "^3.5.3",
-        "prettier-plugin-tailwindcss": "^0.6.11",
-        "typescript": "^5.8.3",
-        "typescript-eslint": "^8.31.0",
-        "vite": "^6.3.3"
-    },
-    "dependencies": {
-        "@tailwindcss/vite": "^4.1.4",
-        "tailwindcss": "^4.1.4",
-        "webextension-polyfill": "^0.12.0"
-    },
-    "bunrc": {
-        "run": {
-            "bun": true
-        }
-    }
-}
