--- web-extension/popup.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <link rel="icon" type="image" href="/favicon.png" />
-        <title>Patch Roulette Extension Popup</title>
-        <script type="module" src="/src/popup.ts"></script>
-        <link rel="stylesheet" href="/src/style.css" />
-    </head>
-    <body class="w-fit p-3 text-base">
-        <div class="flex flex-col gap-1.5 text-nowrap">
-            <button id="open-button" type="button" class="rounded bg-blue-500 px-4 py-2 text-white hover:bg-blue-700">
-                Open in Patch Roulette Diff Viewer
-            </button>
-            <button id="settings-button" type="button" class="rounded bg-blue-500 px-4 py-2 text-white hover:bg-blue-700">Extension Options</button>
-        </div>
-    </body>
-</html>
