--- web-extension/src/style.css
@@ -1,1 +1,0 @@
-@import "tailwindcss";
